+++ src/InvoiceData/Settings/ConfigProxy.php
@@ -7,7 +7,7 @@
 
 declare(strict_types=1);
 
-namespace FreshAdvance\Invoice\Settings;
+namespace FreshAdvance\Invoice\InvoiceData\Settings;
 
 use OxidEsales\Eshop\Core\Config;