+++ src/InvoiceData/Settings/ModuleSettingsInterface.php
@@ -7,12 +7,13 @@
 
 declare(strict_types=1);
 
-namespace FreshAdvance\Invoice\Settings;
+namespace FreshAdvance\Invoice\InvoiceData\Settings;
 
 interface ModuleSettingsInterface
 {
     public function getFileNameFormat(): string;
 
+    // todo: split into separate interface, it should be in Pdf namespace
     public function isForArchive(): bool;
 
     public function getInvoiceNumberFormat(): string;